@@ -1,5 +1,5 @@
 {
-  "name": "elish-master",
+  "name": "e-store",
   "version": "0.0.0",
   "scripts": {
     "ng": "ng",
@@ -11,49 +11,44 @@
   },
   "private": true,
   "dependencies": {
-    "@angular/animations": "^6.1.0",
-    "@angular/cdk": "^6.4.1",
-    "@angular/common": "^6.1.0",
-    "@angular/compiler": "^6.1.0",
-    "@angular/core": "^6.1.0",
-    "@angular/forms": "^6.1.0",
-    "@angular/http": "^6.1.0",
-    "@angular/material": "^6.4.1",
-    "@angular/platform-browser": "^6.1.0",
-    "@angular/platform-browser-dynamic": "^6.1.0",
-    "@angular/router": "^6.1.0",
-    "@capacitor/cli": "^1.0.0-beta.3",
-    "@capacitor/core": "^1.0.0-beta.3",
-    "angularfire2": "^5.0.0-rc.11",
-    "cordova-plugin-browsertab": "^0.2.0",
-    "cordova-plugin-buildinfo": "^2.0.1",
-    "cordova-plugin-inappbrowser": "^3.0.0",
-    "cordova-universal-links-plugin": "^1.2.1",
-    "core-js": "^2.5.7",
-    "firebase": "^5.3.1",
-    "rxjs": "^6.2.2",
-    "rxjs-compat": "^6.2.2",
-    "zone.js": "^0.8.26"
+    "@angular/animations": "^8.2.1",
+    "@angular/cdk": "^8.1.2",
+    "@angular/common": "~8.2.0",
+    "@angular/compiler": "~8.2.0",
+    "@angular/core": "~8.2.0",
+    "@angular/fire": "^5.2.1",
+    "@angular/forms": "~8.2.0",
+    "@angular/http": "^7.2.15",
+    "@angular/material": "^8.1.2",
+    "@angular/platform-browser": "~8.2.0",
+    "@angular/platform-browser-dynamic": "~8.2.0",
+    "@angular/router": "~8.2.0",
+    "firebase": "^6.3.5",
+    "hammerjs": "^2.0.8",
+    "rxjs": "~6.4.0",
+    "rxjs-compat": "^6.5.2",
+    "tslib": "^1.10.0",
+    "zone.js": "~0.9.1"
   },
   "devDependencies": {
-    "@angular-devkit/build-angular": "^0.6.8",
-    "@angular/cli": "^6.1.1",
-    "@angular/compiler-cli": "^6.1.0",
-    "@angular/language-service": "^6.1.0",
-    "@types/jasmine": "^2.8.8",
-    "@types/jasminewd2": "~2.0.3",
+    "@angular-devkit/build-angular": "~0.802.1",
+    "@angular/cli": "~8.2.1",
+    "@angular/compiler-cli": "~8.2.0",
+    "@angular/language-service": "~8.2.0",
     "@types/node": "~8.9.4",
-    "codelyzer": "~4.2.1",
-    "jasmine-core": "~2.99.1",
+    "@types/jasmine": "~3.3.8",
+    "@types/jasminewd2": "~2.0.3",
+    "codelyzer": "^5.0.0",
+    "jasmine-core": "~3.4.0",
     "jasmine-spec-reporter": "~4.2.1",
-    "karma": "~1.7.1",
+    "karma": "~4.1.0",
     "karma-chrome-launcher": "~2.2.0",
-    "karma-coverage-istanbul-reporter": "~1.4.2",
-    "karma-jasmine": "~1.1.1",
-    "karma-jasmine-html-reporter": "^0.2.2",
-    "protractor": "~5.3.0",
-    "ts-node": "~5.0.1",
-    "tslint": "~5.9.1",
-    "typescript": "~2.7.2"
+    "karma-coverage-istanbul-reporter": "~2.0.1",
+    "karma-jasmine": "~2.0.1",
+    "karma-jasmine-html-reporter": "^1.4.0",
+    "protractor": "~5.4.0",
+    "ts-node": "~7.0.0",
+    "tslint": "~5.15.0",
+    "typescript": "~3.5.3"
   }
 }
